chatboat/requirements.txt
ruff
fastapi
uvicorn
python-dotenv
openai
uvicorn
